+++ BH/Modules/Info/RecipeTab.h
@@ -83,7 +83,7 @@ class RecipeTab : public InfoTab {
 		RecipeTab(Drawing::UI* ui);
 
 		void MpqLoaded();
-		bool HandlesCommand(const std::string& command);
+		std::vector<ChatCommand> GetCommands();
 		void OnDraw();
 		bool OnKey(bool up, BYTE key);
 		void OnOpen();