+++ BH/Modules/Info/SetTab.h
@@ -108,7 +108,7 @@ class SetTab : public InfoTab {
 		SetTab(Drawing::UI* ui);
 
 		void MpqLoaded();
-		bool HandlesCommand(const std::string& command);
+		std::vector<ChatCommand> GetCommands();
 		void OnDraw();
 		bool OnKey(bool up, BYTE key);
 		void OnOpen();